+++ src/App.js
@@ -13,8 +13,10 @@ import Cookies from 'cookies.js';
 import Login from './pages/Login/Login';
 import Portal from './pages/Portal/Portal';
 import Dashboard from './pages/Dashboard/Dashboard';
+import Settings from './pages/Settings/Settings';
 import ProfileDashboard from './pages/ProfileDashboard/ProfileDashboard';
 import MinderGasNlSettings from './pages/MinderGasNlSettings/MinderGasNlSettings';
+import WaterReaderSettings from './pages/WaterReaderSettings/WaterReaderSettings';
 import FourOFour from './pages/FourOFour/FourOFour';
 
 // TODO: doesn't work after logging in
@@ -53,11 +55,17 @@ export default function App() {
 					<PrivateRoute redirectComponents={<Portal />} exact path="/">
 						<Dashboard />
 					</PrivateRoute>
+					<PrivateRoute path="/settings">
+						<Settings />
+					</PrivateRoute>
 					<PrivateRoute path="/profile">
 						<ProfileDashboard />
 					</PrivateRoute>
 					<PrivateRoute path="/minder-gas-nl">
 						<MinderGasNlSettings />
+					</PrivateRoute>
+					<PrivateRoute path="/water-reader">
+						<WaterReaderSettings />
 					</PrivateRoute>
 					<Route>
 						<FourOFour />