+++ src/atoms/ArrowLeftIcon/ArrowLeftIcon.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function ArrowLeftIcon() {
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 18 18"><path fill="currentColor" d="M9,18l1.59-1.59L4.31,10.12H18V7.88H4.31l6.28-6.28L9,0,0,9Z"/></svg>
+	);
+}