+++ src/molecules/PageNav/PageNav.js
@@ -1,0 +1,23 @@
+import styles from './PageNav.module.css';
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+export default function PageNav({ menuItems = [] }) {
+	return (
+		<nav className={styles.container}>
+			<ul>
+				{menuItems.map(item => (
+					<li>
+						<Link to={item.to}>
+							<div className={styles.icon}>
+								<item.icon />
+							</div>
+							<span className={styles.title}>{item.label}</span>
+							<span className={styles.description}>{item.description}</span>
+						</Link>
+					</li>
+				))}
+			</ul>
+		</nav>
+	);
+}