+++ src/organisms/Header/Header.js
@@ -3,13 +3,17 @@ import styles from './Header.module.css';
 import { Link } from 'react-router-dom';
 
 import UserMenu from '../../organisms/UserMenu/UserMenu';
-// import Navigation from '../Navigation/Navigation';
+
+import CogwheelIcon from '../../atoms/CogwheelIcon/CogwheelIcon';
 
 export default function Header() {
 	return (
 		<header className={styles.container}>
 			<div className={styles.main}>
-				<Link to="/" className={styles.appName}>Home Care</Link>
+				<div>
+					<Link to="/" className={styles.appName}>Home Care</Link>
+				</div>
+				<Link to="/settings"><CogwheelIcon /></Link>
 				<UserMenu className={styles.userMenu} />
 			</div>
 			{/* <div className="header-navigation-container">