+++ CHANGELOG.md
@@ -4,4 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2021-04-12
+### Added
+- Settings page.
+- Water Reader settings page.
+- EventList loading indication.
+
+### Changed
+- Moved minder-gas-nl page to settings menu.
+
 ## [1.0.0] - 2021-04-10